More organization and add apple sleep queries
