+++ metabase-api/init/questions/utils.py
@@ -4,11 +4,6 @@ import uuid
 from filters.utils import add_field_filters, get_field_mappings
 from loguru import logger
 from metabase_api import Metabase_API
-from queries.apple_health import query_calories_burned
-from queries.strong import (query_count_by_workout_type,
-                            query_duration_by_workout_type,
-                            query_sets_by_workout_type,
-                            query_volume_by_exercise_type)
 
 
 def set_visualization_settings(show_values: bool = True, x_axis_title: str = None, y_axis_title: str = None, dimensions: list = None, metrics: list = None):
@@ -136,70 +131,3 @@ def create_sql_timeseries_question(mb: Metabase_API, query: str, display: str = 
         logger.success(f"Successfully created question - {question_name}")
     except Exception as e:
         logger.error(f"Could not create question - {question_name}\n{e}")
-
-############################################
-# Strong App Questions
-############################################
-
-
-def strong_workout_duration_by_type(mb: Metabase_API):
-    query = query_duration_by_workout_type()
-    visualization_settings = set_visualization_settings(
-        x_axis_title="Workout Type",
-        y_axis_title="Avg. Workout Duration (min)",
-        dimensions=["workout_name"],
-        metrics=["average_workout_length_minutes",
-                 "median_workout_length_minutes"]
-    )
-    create_sql_question(mb, query=query, question_name="Workout Duration by Type",
-                        display="bar", db_id=2, collection_id=2, table_id=48, visualization_settings=visualization_settings, timestamp_field_name="created_at")
-
-
-def strong_sets_by_workout_type(mb: Metabase_API):
-    query = query_sets_by_workout_type()
-    visualization_settings = set_visualization_settings(
-        x_axis_title="Time Period",
-        y_axis_title="Number of Sets",
-        dimensions=["time_period", "workout_name"],
-        metrics=["number_of_sets"]
-    )
-    create_sql_timeseries_question(mb, query=query, question_name="Sets Over Time",
-                                   display="line", db_id=2, collection_id=2, table_id=48, visualization_settings=visualization_settings)
-
-
-def strong_volume_by_exercise_type(mb: Metabase_API):
-    query = query_volume_by_exercise_type()
-    visualization_settings = set_visualization_settings(
-        dimensions=["time_period", "workout_name", "exercise_name"],
-        metrics=["time_period", "workout_name", "exercise_name"]
-    )
-    create_sql_timeseries_question(mb, query=query, question_name="Exercises by Volume",
-                                   display="table", db_id=2, collection_id=2, table_id=48, visualization_settings=visualization_settings)
-
-
-def strong_count_by_workout_type(mb: Metabase_API):
-    query = query_count_by_workout_type()
-    visualization_settings = set_visualization_settings(
-        x_axis_title="Time Period",
-        y_axis_title="Number of Sets",
-        dimensions=["time_period", "workout_name"],
-        metrics=["number_of_workout_days"]
-    )
-    create_sql_timeseries_question(mb, query=query, question_name="Workouts Over Time",
-                                   display="line", db_id=2, collection_id=2, table_id=48, visualization_settings=visualization_settings)
-
-
-############################################
-# Apple Health Questions
-############################################
-def apple_calories(mb: Metabase_API):
-    query = query_calories_burned()
-    visualization_settings = set_visualization_settings(
-        x_axis_title="Time Period",
-        y_axis_title="Calories Burned",
-        dimensions=["time_period"],
-        metrics=["total_calories_burned",
-                 "active_calories_burned", "resting_calories_burned"]
-    )
-    create_sql_timeseries_question(mb, query=query, question_name="Calories Burned Over Time",
-                                   display="line", db_id=2, collection_id=3, table_id=40, visualization_settings=visualization_settings)
