+++ metabase-api/main.py
@@ -2,10 +2,12 @@ import os
 
 from init.auth import auth
 from init.collections import create_collection
-from init.questions import (apple_calories, strong_count_by_workout_type,
-                            strong_sets_by_workout_type,
-                            strong_volume_by_exercise_type,
-                            strong_workout_duration_by_type)
+from init.questions.apple import (apple_calories, apple_rem_cycles,
+                                  apple_sleep_hours)
+from init.questions.strong import (strong_count_by_workout_type,
+                                   strong_sets_by_workout_type,
+                                   strong_volume_by_exercise_type,
+                                   strong_workout_duration_by_type)
 from loguru import logger
 
 # Load environment variables from the .env file
@@ -34,6 +36,8 @@ if mb:
 
     # Inside the Apple Health Collection
     apple_calories(mb)
+    apple_sleep_hours(mb)
+    apple_rem_cycles(mb)
 
     # Success message
     logger.success("Successfully initialized all metabase questions")
