+++ metabase-api/queries/apple/activity.py
@@ -1,7 +1,7 @@
 # Description: Metabase queries for the Apple Health Data
 
-def add_apple_field_filters_to_sql(query: str) -> str:
-    """Replaces generic `where 1=1` with Metabase field filters for the Strong App.
+def add_apple_activity_field_filters_to_sql(query: str) -> str:
+    """Replaces generic `where 1=1` with Metabase field filters for the Apple Health Activity Data.
 
     Args:
         query (str): The SQL query to be modified.
@@ -19,10 +19,10 @@ def query_calories_burned():
     , sum(active_energy_burned) as active_calories_burned
     , sum(basal_energy_burned) as resting_calories_burned
     , sum(active_energy_burned) + sum(basal_energy_burned) as total_calories_burned
-from apple_health_raw
+from apple_health_activity_raw
 where 1=1
 group by time_period
 order by time_period desc
 """
-    query = add_apple_field_filters_to_sql(query)
+    query = add_apple_activity_field_filters_to_sql(query)
     return query.strip()
