+++ metabase-api/queries/strong/lifting.py
